Refactor meetup settings page
